JAM-232 #time 20m changing makeJsonFrontentCompatible and UserFlowchartRepo to align with new flowchart and user models
